fix: exception hierarchy, move ApprovalTransition to own file, remove dead code, fix KDoc
