+++ tramai-core/src/main/kotlin/dev/tramai/core/approval/IllegalApprovalTransitionException.kt
@@ -1,10 +1,12 @@
 package dev.tramai.core.approval
 
+import dev.tramai.core.exception.TramaiException
+
 class IllegalApprovalTransitionException(
     val approvalId: String,
     val from: ApprovalStatus,
     val to: ApprovalStatus,
     val reason: String,
-) : IllegalArgumentException(
+) : TramaiException(
     "Illegal approval transition for '$approvalId': $from -> $to - $reason"
 )
